Refactored prefix search and made Trie api more sane
Prefix search is now a lot easier to understand as it just recursively calls find.  Also added the option to return the last node of a find even if there was no match on the node in order to enable this.  I tried to make the Trie functions operate more like Map.get/1 and Map.put/2 as much as I could.  Removed all of the genserver code as it's not really a concern of this library.  In the application that uses it we can use the appropriate OTP setup.